requirements.txt
Flask
numpy
scikit-learn
annoy
networkx
requests
chromadb
python-dotenv